Fix: untangle a mixture of web browsers and devices; it is all devices now
