+++ src/php/DataSift/Storyplayer/DeviceLib/E5xx/BadDeviceAdapter.php
@@ -34,32 +34,32 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/WebBrowserLib
+ * @package   Storyplayer/DeviceLib
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\WebBrowserLib;
+namespace DataSift\Storyplayer\DeviceLib;
 
 use DataSift\Stone\ExceptionsLib\Exxx_Exception;
 
 /**
- * Exception thrown when we've loaded a web browser adapter, but it doesn't
- * implement the WebBrowserAdapter interface
+ * Exception thrown when we've loaded a device adapter, but it doesn't
+ * implement the DeviceAdapter interface
  *
  * @category    Libraries
- * @package     Storyplayer/WebBrowserLib
+ * @package     Storyplayer/DeviceLib
  * @author      Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright   2011-present Mediasift Ltd www.datasift.com
  * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link        http://datasift.github.io/storyplayer
  */
-class E5xx_BadWebBrowserAdapter extends Exxx_Exception
+class E5xx_BadDeviceAdapter extends Exxx_Exception
 {
 	public function __construct($classname) {
-		$msg = "Unable to use class '{$classname}' as a web browser adapter; it does not implement the WebBrowserAdapter interface";
+		$msg = "Unable to use class '{$classname}' as a device adapter; it does not implement the DeviceAdapter interface";
 		parent::__construct(500, $msg, $msg);
 	}
 }
