+++ src/php/DataSift/Storyplayer/DeviceLib/LocalWebDriverAdapter.php
@@ -34,31 +34,86 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/PlayerLib
+ * @package   Storyplayer/DeviceLib
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\PlayerLib;
+namespace DataSift\Storyplayer\DeviceLib;
 
-use DataSift\Stone\ExceptionsLib\Exxx_Exception;
+use Exception;
+use DataSift\BrowserMobProxy\BrowserMobProxyClient;
+use DataSift\Storyplayer\PlayerLib\StoryTeller;
+use DataSift\WebDriver\WebDriverClient;
 
 /**
- * Exception thrown when we've tried, but failed, to start the web browser
+ * The adapter that talks to Browsermob-proxy and Selenium-standalone-server
+ * running on the same host as Storyplayer
  *
  * @category    Libraries
- * @package     Storyplayer/PlayerLib
+ * @package     Storyplayer/DeviceLib
  * @author      Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright   2011-present Mediasift Ltd www.datasift.com
  * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link        http://datasift.github.io/storyplayer
  */
-class E5xx_CannotStartWebBrowser extends Exxx_Exception
+class LocalWebDriverAdapter extends BaseAdapter implements DeviceAdapter
 {
-	public function __construct() {
-		$msg = "Cannot start the web browser; are both browsermob-proxy and selenium running?";
-		parent::__construct(500, $msg, $msg);
+	public function start(StoryTeller $st)
+	{
+		$httpProxy = new BrowserMobProxyClient();
+		$httpProxy->enableFeature('paramLogs');
+
+		$this->proxySession = $httpProxy->createProxy();
+
+		// start recording
+		$this->proxySession->startHAR();
+
+		// create the browser session
+		$webDriver = new WebDriverClient();
+		$this->browserSession = $webDriver->newSession(
+			$this->browserDetails->browser,
+			array(
+				'proxy' => $this->proxySession->getWebDriverProxyConfig()
+			) + $this->browserDetails->desiredCapabilities
+
+		);
+	}
+
+	public function stop()
+	{
+		// stop the web browser
+		if (is_object($this->browserSession))
+		{
+			$this->browserSession->close();
+			$this->browserSession = null;
+		}
+
+		// now stop the proxy
+		if (is_object($this->proxySession))
+		{
+			try {
+				$this->proxySession->close();
+			}
+			catch (Exception $e) {
+				// do nothing - we don't care!
+			}
+			$this->proxySession = null;
+		}
+	}
+
+	public function applyHttpBasicAuthForHost($hostname, $url)
+	{
+		// get the auth credentials
+		$credentials = $this->getHttpBasicAuthForHost($hostname);
+
+		if (isset($this->proxySession)) {
+			$this->proxySession->setHttpBasicAuth($hostname, $credentials['user'], $credentials['pass']);
+		}
+
+		// all done
+		return $url;
 	}
 }
