+++ src/php/DataSift/Storyplayer/DeviceLib/RemoteWebDriverAdapter.php
@@ -34,31 +34,50 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/WebBrowserLib
+ * @package   Storyplayer/DeviceLib
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\WebBrowserLib;
+namespace DataSift\Storyplayer\DeviceLib;
 
-use DataSift\Stone\ExceptionsLib\Exxx_Exception;
+use Exception;
+use DataSift\BrowserMobProxy\BrowserMobProxyClient;
+use DataSift\Storyplayer\PlayerLib\StoryTeller;
+use DataSift\WebDriver\WebDriverClient;
 
 /**
- * Exception thrown when we've tried, but failed, to start the web browser
+ * The adapter that talks to Browsermob-proxy running locally, and a
+ * Selenium Server that is running in an arbitrary remote location
  *
  * @category    Libraries
- * @package     Storyplayer/WebBrowserLib
+ * @package     Storyplayer/DeviceLib
  * @author      Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright   2011-present Mediasift Ltd www.datasift.com
  * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link        http://datasift.github.io/storyplayer
  */
-class E5xx_CannotStartWebBrowser extends Exxx_Exception
+class RemoteWebDriverAdapter extends LocalWebDriverAdapter
 {
-	public function __construct() {
-		$msg = "Cannot start the web browser; are both browsermob-proxy and selenium running?";
-		parent::__construct(500, $msg, $msg);
+	public function start(StoryTeller $st)
+	{
+		$httpProxy = new BrowserMobProxyClient();
+		$httpProxy->enableFeature('paramLogs');
+
+		$this->proxySession = $httpProxy->createProxy();
+
+		// start recording
+		$this->proxySession->startHAR();
+
+		// create the browser session
+		$webDriver = new WebDriverClient($this->browserDetails->url);
+		$this->browserSession = $webDriver->newSession(
+			$this->browserDetails->browser,
+			array(
+				'proxy' => $this->proxySession->getWebDriverProxyConfig()
+			) + $this->browserDetails->desiredCapabilities
+		);
 	}
 }
